backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
biopython==1.83
pandas==2.2.0
numpy==1.26.4
